Enum atsamd51j19a::adc0::inputctrl::MUXPOSR

pub enum MUXPOSR {
    AIN0,
    AIN1,
    AIN2,
    AIN3,
    AIN4,
    AIN5,
    AIN6,
    AIN7,
    AIN8,
    AIN9,
    AIN10,
    AIN11,
    AIN12,
    AIN13,
    AIN14,
    AIN15,
    AIN16,
    AIN17,
    AIN18,
    AIN19,
    AIN20,
    AIN21,
    AIN22,
    AIN23,
    SCALEDCOREVCC,
    SCALEDVBAT,
    SCALEDIOVCC,
    BANDGAP,
    PTAT,
    CTAT,
    DAC,
    PTC,
}

Possible values of the field MUXPOS

Variants
AIN0

ADC AIN0 Pin

AIN1

ADC AIN1 Pin

AIN2

ADC AIN2 Pin

AIN3

ADC AIN3 Pin

AIN4

ADC AIN4 Pin

AIN5

ADC AIN5 Pin

AIN6

ADC AIN6 Pin

AIN7

ADC AIN7 Pin

AIN8

ADC AIN8 Pin

AIN9

ADC AIN9 Pin

AIN10

ADC AIN10 Pin

AIN11

ADC AIN11 Pin

AIN12

ADC AIN12 Pin

AIN13

ADC AIN13 Pin

AIN14

ADC AIN14 Pin

AIN15

ADC AIN15 Pin

AIN16

ADC AIN16 Pin

AIN17

ADC AIN17 Pin

AIN18

ADC AIN18 Pin

AIN19

ADC AIN19 Pin

AIN20

ADC AIN20 Pin

AIN21

ADC AIN21 Pin

AIN22

ADC AIN22 Pin

AIN23

ADC AIN23 Pin

SCALEDCOREVCC

1/4 Scaled Core Supply

SCALEDVBAT

1/4 Scaled VBAT Supply

SCALEDIOVCC

1/4 Scaled I/O Supply

BANDGAP

Bandgap Voltage

PTAT

Temperature Sensor

CTAT

Temperature Sensor

DAC

DAC Output

PTC

PTC output (only on ADC0)

Methods

impl MUXPOSR

pub fn bits(&self) -> u8

Value of the field as raw bits

pub fn is_ain0(&self) -> bool

Checks if the value of the field is AIN0

pub fn is_ain1(&self) -> bool

Checks if the value of the field is AIN1

pub fn is_ain2(&self) -> bool

Checks if the value of the field is AIN2

pub fn is_ain3(&self) -> bool

Checks if the value of the field is AIN3

pub fn is_ain4(&self) -> bool

Checks if the value of the field is AIN4

pub fn is_ain5(&self) -> bool

Checks if the value of the field is AIN5

pub fn is_ain6(&self) -> bool

Checks if the value of the field is AIN6

pub fn is_ain7(&self) -> bool

Checks if the value of the field is AIN7

pub fn is_ain8(&self) -> bool

Checks if the value of the field is AIN8

pub fn is_ain9(&self) -> bool

Checks if the value of the field is AIN9

pub fn is_ain10(&self) -> bool

Checks if the value of the field is AIN10

pub fn is_ain11(&self) -> bool

Checks if the value of the field is AIN11

pub fn is_ain12(&self) -> bool

Checks if the value of the field is AIN12

pub fn is_ain13(&self) -> bool

Checks if the value of the field is AIN13

pub fn is_ain14(&self) -> bool

Checks if the value of the field is AIN14

pub fn is_ain15(&self) -> bool

Checks if the value of the field is AIN15

pub fn is_ain16(&self) -> bool

Checks if the value of the field is AIN16

pub fn is_ain17(&self) -> bool

Checks if the value of the field is AIN17

pub fn is_ain18(&self) -> bool

Checks if the value of the field is AIN18

pub fn is_ain19(&self) -> bool

Checks if the value of the field is AIN19

pub fn is_ain20(&self) -> bool

Checks if the value of the field is AIN20

pub fn is_ain21(&self) -> bool

Checks if the value of the field is AIN21

pub fn is_ain22(&self) -> bool

Checks if the value of the field is AIN22

pub fn is_ain23(&self) -> bool

Checks if the value of the field is AIN23

pub fn is_scaledcorevcc(&self) -> bool

Checks if the value of the field is SCALEDCOREVCC

pub fn is_scaledvbat(&self) -> bool

Checks if the value of the field is SCALEDVBAT

pub fn is_scalediovcc(&self) -> bool

Checks if the value of the field is SCALEDIOVCC

pub fn is_bandgap(&self) -> bool

Checks if the value of the field is BANDGAP

pub fn is_ptat(&self) -> bool

Checks if the value of the field is PTAT

pub fn is_ctat(&self) -> bool

Checks if the value of the field is CTAT

pub fn is_dac(&self) -> bool

Checks if the value of the field is DAC

pub fn is_ptc(&self) -> bool

Checks if the value of the field is PTC