Module atsamd51j19a::aes::ctrla

Control A

Structs

CTYPER
Value of the field

ENABLER
Value of the field

R
Value read from the register

SWRSTR
Value of the field

W
Value to write to the register

_AESMODEW
Proxy

_CFBSW
Proxy

_CIPHERW
Proxy

_CTYPEW
Proxy

_ENABLEW
Proxy

_KEYGENW
Proxy

_KEYSIZEW
Proxy

_LODW
Proxy

_STARTMODEW
Proxy

_SWRSTW
Proxy

_XORKEYW
Proxy

Enums

AESMODER
Possible values of the field AESMODE

AESMODEW
Values that can be written to the field AESMODE

CFBSR
Possible values of the field CFBS

CFBSW
Values that can be written to the field CFBS

CIPHERR
Possible values of the field CIPHER

CIPHERW
Values that can be written to the field CIPHER

KEYGENR
Possible values of the field KEYGEN

KEYGENW
Values that can be written to the field KEYGEN

KEYSIZER
Possible values of the field KEYSIZE

KEYSIZEW
Values that can be written to the field KEYSIZE

LODR
Possible values of the field LOD

LODW
Values that can be written to the field LOD

STARTMODER
Possible values of the field STARTMODE

STARTMODEW
Values that can be written to the field STARTMODE

XORKEYR
Possible values of the field XORKEY

XORKEYW
Values that can be written to the field XORKEY