Enum atsamd51j19a::dmac::chctrla::BURSTLENR

pub enum BURSTLENR {
    SINGLE,
    _2BEAT,
    _3BEAT,
    _4BEAT,
    _5BEAT,
    _6BEAT,
    _7BEAT,
    _8BEAT,
    _9BEAT,
    _10BEAT,
    _11BEAT,
    _12BEAT,
    _13BEAT,
    _14BEAT,
    _15BEAT,
    _16BEAT,
}

Possible values of the field BURSTLEN

Variants
SINGLE

Single-beat burst length

_2BEAT

2-beats burst length

_3BEAT

3-beats burst length

_4BEAT

4-beats burst length

_5BEAT

5-beats burst length

_6BEAT

6-beats burst length

_7BEAT

7-beats burst length

_8BEAT

8-beats burst length

_9BEAT

9-beats burst length

_10BEAT

10-beats burst length

_11BEAT

11-beats burst length

_12BEAT

12-beats burst length

_13BEAT

13-beats burst length

_14BEAT

14-beats burst length

_15BEAT

15-beats burst length

_16BEAT

16-beats burst length

Methods

impl BURSTLENR

pub fn bits(&self) -> u8

Value of the field as raw bits

pub fn is_single(&self) -> bool

Checks if the value of the field is SINGLE

pub fn is_2beat(&self) -> bool

Checks if the value of the field is _2BEAT

pub fn is_3beat(&self) -> bool

Checks if the value of the field is _3BEAT

pub fn is_4beat(&self) -> bool

Checks if the value of the field is _4BEAT

pub fn is_5beat(&self) -> bool

Checks if the value of the field is _5BEAT

pub fn is_6beat(&self) -> bool

Checks if the value of the field is _6BEAT

pub fn is_7beat(&self) -> bool

Checks if the value of the field is _7BEAT

pub fn is_8beat(&self) -> bool

Checks if the value of the field is _8BEAT

pub fn is_9beat(&self) -> bool

Checks if the value of the field is _9BEAT

pub fn is_10beat(&self) -> bool

Checks if the value of the field is _10BEAT

pub fn is_11beat(&self) -> bool

Checks if the value of the field is _11BEAT

pub fn is_12beat(&self) -> bool

Checks if the value of the field is _12BEAT

pub fn is_13beat(&self) -> bool

Checks if the value of the field is _13BEAT

pub fn is_14beat(&self) -> bool

Checks if the value of the field is _14BEAT

pub fn is_15beat(&self) -> bool

Checks if the value of the field is _15BEAT

pub fn is_16beat(&self) -> bool

Checks if the value of the field is _16BEAT