Enum atsamd51j19a::gclk::syncbusy::GENCTRL8R

pub enum GENCTRL8R {
    GCLK0,
    GCLK1,
    GCLK2,
    GCLK3,
    GCLK4,
    GCLK5,
    GCLK6,
    GCLK7,
    GCLK8,
    GCLK9,
    GCLK10,
    GCLK11,
    _Reserved(bool),
}

Possible values of the field GENCTRL8

Variants
GCLK0

Generic clock generator 0

GCLK1

Generic clock generator 1

GCLK2

Generic clock generator 2

GCLK3

Generic clock generator 3

GCLK4

Generic clock generator 4

GCLK5

Generic clock generator 5

GCLK6

Generic clock generator 6

GCLK7

Generic clock generator 7

GCLK8

Generic clock generator 8

GCLK9

Generic clock generator 9

GCLK10

Generic clock generator 10

GCLK11

Generic clock generator 11

_Reserved(bool)

Reserved

Methods

impl GENCTRL8R

pub fn bit_is_clear(&self) -> bool

Returns true if the bit is clear (0)

pub fn bit_is_set(&self) -> bool

Returns true if the bit is set (1)

pub fn bit(&self) -> bool

Value of the field as raw bits

pub fn is_gclk0(&self) -> bool

Checks if the value of the field is GCLK0

pub fn is_gclk1(&self) -> bool

Checks if the value of the field is GCLK1

pub fn is_gclk2(&self) -> bool

Checks if the value of the field is GCLK2

pub fn is_gclk3(&self) -> bool

Checks if the value of the field is GCLK3

pub fn is_gclk4(&self) -> bool

Checks if the value of the field is GCLK4

pub fn is_gclk5(&self) -> bool

Checks if the value of the field is GCLK5

pub fn is_gclk6(&self) -> bool

Checks if the value of the field is GCLK6

pub fn is_gclk7(&self) -> bool

Checks if the value of the field is GCLK7

pub fn is_gclk8(&self) -> bool

Checks if the value of the field is GCLK8

pub fn is_gclk9(&self) -> bool

Checks if the value of the field is GCLK9

pub fn is_gclk10(&self) -> bool

Checks if the value of the field is GCLK10

pub fn is_gclk11(&self) -> bool

Checks if the value of the field is GCLK11