Module atsamd51j19a::i2s::txctrl

Tx Serializer Control

Structs

R
Value read from the register

SLOTDIS0R
Value of the field

SLOTDIS1R
Value of the field

SLOTDIS2R
Value of the field

SLOTDIS3R
Value of the field

SLOTDIS4R
Value of the field

SLOTDIS5R
Value of the field

SLOTDIS6R
Value of the field

SLOTDIS7R
Value of the field

W
Value to write to the register

_BITREVW
Proxy

_DATASIZEW
Proxy

_DMAW
Proxy

_EXTENDW
Proxy

_MONOW
Proxy

_SLOTADJW
Proxy

_SLOTDIS0W
Proxy

_SLOTDIS1W
Proxy

_SLOTDIS2W
Proxy

_SLOTDIS3W
Proxy

_SLOTDIS4W
Proxy

_SLOTDIS5W
Proxy

_SLOTDIS6W
Proxy

_SLOTDIS7W
Proxy

_TXDEFAULTW
Proxy

_TXSAMEW
Proxy

_WORDADJW
Proxy

Enums

BITREVR
Possible values of the field BITREV

BITREVW
Values that can be written to the field BITREV

DATASIZER
Possible values of the field DATASIZE

DATASIZEW
Values that can be written to the field DATASIZE

DMAR
Possible values of the field DMA

DMAW
Values that can be written to the field DMA

EXTENDR
Possible values of the field EXTEND

EXTENDW
Values that can be written to the field EXTEND

MONOR
Possible values of the field MONO

MONOW
Values that can be written to the field MONO

SLOTADJR
Possible values of the field SLOTADJ

SLOTADJW
Values that can be written to the field SLOTADJ

TXDEFAULTR
Possible values of the field TXDEFAULT

TXDEFAULTW
Values that can be written to the field TXDEFAULT

TXSAMER
Possible values of the field TXSAME

TXSAMEW
Values that can be written to the field TXSAME

WORDADJR
Possible values of the field WORDADJ

WORDADJW
Values that can be written to the field WORDADJ