Module atsamd51j19a::nvmctrl::ctrla

Control A

Structs

AHBNS0R
Value of the field

AHBNS1R
Value of the field

AUTOWSR
Value of the field

CACHEDIS0R
Value of the field

CACHEDIS1R
Value of the field

R
Value read from the register

RWSR
Value of the field

SUSPENR
Value of the field

W
Value to write to the register

_AHBNS0W
Proxy

_AHBNS1W
Proxy

_AUTOWSW
Proxy

_CACHEDIS0W
Proxy

_CACHEDIS1W
Proxy

_PRMW
Proxy

_RWSW
Proxy

_SUSPENW
Proxy

_WMODEW
Proxy

Enums

PRMR
Possible values of the field PRM

PRMW
Values that can be written to the field PRM

WMODER
Possible values of the field WMODE

WMODEW
Values that can be written to the field WMODE