Enum atsamd51j19a::nvmctrl::ctrlb::CMDW

pub enum CMDW {
    EP,
    EB,
    WP,
    WQW,
    SWRST,
    LR,
    UR,
    SPRM,
    CPRM,
    PBC,
    SSB,
    BKSWRST,
    CELCK,
    CEULCK,
    SBPDIS,
    CBPDIS,
    ASEES0,
    ASEES1,
    SEERALOC,
    SEEFLUSH,
    LSEE,
    USEE,
    LSEER,
    USEER,
}

Values that can be written to the field CMD

Variants
EP

Erase Page - Only supported in the USER and AUX pages.

EB

Erase Block - Erases the block addressed by the ADDR register, not supported in the user page

WP

Write Page - Writes the contents of the page buffer to the page addressed by the ADDR register, not supported in the user page

WQW

Write Quad Word - Writes a 128-bit word at the location addressed by the ADDR register.

SWRST

Software Reset - Power-Cycle the NVM memory and replay the device automatic calibration procedure and resets the module configuration registers

LR

Lock Region - Locks the region containing the address location in the ADDR register.

UR

Unlock Region - Unlocks the region containing the address location in the ADDR register.

SPRM

Sets the power reduction mode.

CPRM

Clears the power reduction mode.

PBC

Page Buffer Clear - Clears the page buffer.

SSB

Set Security Bit

BKSWRST

Bank swap and system reset, if SMEE is used also reallocate SMEE data into the opposite BANK

CELCK

Chip Erase Lock - DSU.CE command is not available

CEULCK

Chip Erase Unlock - DSU.CE command is available

SBPDIS

Sets STATUS.BPDIS, Boot loader protection is discarded until CBPDIS is issued or next start-up sequence

CBPDIS

Clears STATUS.BPDIS, Boot loader protection is not discarded

ASEES0

Activate SmartEEPROM Sector 0, deactivate Sector 1

ASEES1

Activate SmartEEPROM Sector 1, deactivate Sector 0

SEERALOC

Starts SmartEEPROM sector reallocation algorithm

SEEFLUSH

Flush SMEE data when in buffered mode

LSEE

Lock access to SmartEEPROM data from any mean

USEE

Unlock access to SmartEEPROM data

LSEER

Lock access to the SmartEEPROM Register Address Space (above 64KB)

USEER

Unlock access to the SmartEEPROM Register Address Space (above 64KB)