Module atsamd51j19a::pdec::ctrla

Control A

Structs

ALOCKR
Value of the field

ANGULARR
Value of the field

ENABLER
Value of the field

MAXCMPR
Value of the field

PERENR
Value of the field

PINEN0R
Value of the field

PINEN1R
Value of the field

PINEN2R
Value of the field

PINVEN0R
Value of the field

PINVEN1R
Value of the field

PINVEN2R
Value of the field

R
Value read from the register

RUNSTDBYR
Value of the field

SWAPR
Value of the field

W
Value to write to the register

_ALOCKW
Proxy

_ANGULARW
Proxy

_CONFW
Proxy

_ENABLEW
Proxy

_MAXCMPW
Proxy

_MODEW
Proxy

_PERENW
Proxy

_PINEN0W
Proxy

_PINEN1W
Proxy

_PINEN2W
Proxy

_PINVEN0W
Proxy

_PINVEN1W
Proxy

_PINVEN2W
Proxy

_RUNSTDBYW
Proxy

_SWAPW
Proxy

_SWRSTW
Proxy

Enums

CONFR
Possible values of the field CONF

CONFW
Values that can be written to the field CONF

MODER
Possible values of the field MODE

MODEW
Values that can be written to the field MODE