Module atsamd51j19a::qspi::ctrlb

Control B

Structs

DLYBCTR
Value of the field

DLYCSR
Value of the field

LOOPENR
Value of the field

R
Value read from the register

SMEMREGR
Value of the field

W
Value to write to the register

WDRBTR
Value of the field

_CSMODEW
Proxy

_DATALENW
Proxy

_DLYBCTW
Proxy

_DLYCSW
Proxy

_LOOPENW
Proxy

_MODEW
Proxy

_SMEMREGW
Proxy

_WDRBTW
Proxy

Enums

CSMODER
Possible values of the field CSMODE

CSMODEW
Values that can be written to the field CSMODE

DATALENR
Possible values of the field DATALEN

DATALENW
Values that can be written to the field DATALEN

MODER
Possible values of the field MODE

MODEW
Values that can be written to the field MODE