Enum atsamd51j19a::qspi::ctrlb::DATALENR

pub enum DATALENR {
    _8BITS,
    _9BITS,
    _10BITS,
    _11BITS,
    _12BITS,
    _13BITS,
    _14BITS,
    _15BITS,
    _16BITS,
    _Reserved(u8),
}

Possible values of the field DATALEN

Variants
_8BITS

8-bits transfer

_9BITS

9 bits transfer

_10BITS

10-bits transfer

_11BITS

11-bits transfer

_12BITS

12-bits transfer

_13BITS

13-bits transfer

_14BITS

14-bits transfer

_15BITS

15-bits transfer

_16BITS

16-bits transfer

_Reserved(u8)

Reserved

Methods

impl DATALENR

pub fn bits(&self) -> u8

Value of the field as raw bits

pub fn is_8bits(&self) -> bool

Checks if the value of the field is _8BITS

pub fn is_9bits(&self) -> bool

Checks if the value of the field is _9BITS

pub fn is_10bits(&self) -> bool

Checks if the value of the field is _10BITS

pub fn is_11bits(&self) -> bool

Checks if the value of the field is _11BITS

pub fn is_12bits(&self) -> bool

Checks if the value of the field is _12BITS

pub fn is_13bits(&self) -> bool

Checks if the value of the field is _13BITS

pub fn is_14bits(&self) -> bool

Checks if the value of the field is _14BITS

pub fn is_15bits(&self) -> bool

Checks if the value of the field is _15BITS

pub fn is_16bits(&self) -> bool

Checks if the value of the field is _16BITS