Module atsamd51j19a::qspi::instrframe

Instruction Frame

Structs

ADDRENR
Value of the field

CRMODER
Value of the field

DATAENR
Value of the field

DDRENR
Value of the field

DUMMYLENR
Value of the field

INSTRENR
Value of the field

OPTCODEENR
Value of the field

R
Value read from the register

W
Value to write to the register

_ADDRENW
Proxy

_ADDRLENW
Proxy

_CRMODEW
Proxy

_DATAENW
Proxy

_DDRENW
Proxy

_DUMMYLENW
Proxy

_INSTRENW
Proxy

_OPTCODEENW
Proxy

_OPTCODELENW
Proxy

_TFRTYPEW
Proxy

_WIDTHW
Proxy

Enums

ADDRLENR
Possible values of the field ADDRLEN

ADDRLENW
Values that can be written to the field ADDRLEN

OPTCODELENR
Possible values of the field OPTCODELEN

OPTCODELENW
Values that can be written to the field OPTCODELEN

TFRTYPER
Possible values of the field TFRTYPE

TFRTYPEW
Values that can be written to the field TFRTYPE

WIDTHR
Possible values of the field WIDTH

WIDTHW
Values that can be written to the field WIDTH