Module atsamd51j19a::rtc::mode1::tampctrl

Tamper Control

Structs

DEBNC0R
Value of the field

DEBNC1R
Value of the field

DEBNC2R
Value of the field

DEBNC3R
Value of the field

DEBNC4R
Value of the field

R
Value read from the register

TAMLVL0R
Value of the field

TAMLVL1R
Value of the field

TAMLVL2R
Value of the field

TAMLVL3R
Value of the field

TAMLVL4R
Value of the field

W
Value to write to the register

_DEBNC0W
Proxy

_DEBNC1W
Proxy

_DEBNC2W
Proxy

_DEBNC3W
Proxy

_DEBNC4W
Proxy

_IN0ACTW
Proxy

_IN1ACTW
Proxy

_IN2ACTW
Proxy

_IN3ACTW
Proxy

_IN4ACTW
Proxy

_TAMLVL0W
Proxy

_TAMLVL1W
Proxy

_TAMLVL2W
Proxy

_TAMLVL3W
Proxy

_TAMLVL4W
Proxy

Enums

IN0ACTR
Possible values of the field IN0ACT

IN0ACTW
Values that can be written to the field IN0ACT

IN1ACTR
Possible values of the field IN1ACT

IN1ACTW
Values that can be written to the field IN1ACT

IN2ACTR
Possible values of the field IN2ACT

IN2ACTW
Values that can be written to the field IN2ACT

IN3ACTR
Possible values of the field IN3ACT

IN3ACTW
Values that can be written to the field IN3ACT

IN4ACTR
Possible values of the field IN4ACT

IN4ACTW
Values that can be written to the field IN4ACT