Module atsamd51j19a::rtc::mode2::ctrlb

MODE2 Control B

Structs

DEBASYNCR
Value of the field

DEBMAJR
Value of the field

DMAENR
Value of the field

GP0ENR
Value of the field

GP2ENR
Value of the field

R
Value read from the register

RTCOUTR
Value of the field

W
Value to write to the register

_ACTFW
Proxy

_DEBASYNCW
Proxy

_DEBFW
Proxy

_DEBMAJW
Proxy

_DMAENW
Proxy

_GP0ENW
Proxy

_GP2ENW
Proxy

_RTCOUTW
Proxy

Enums

ACTFR
Possible values of the field ACTF

ACTFW
Values that can be written to the field ACTF

DEBFR
Possible values of the field DEBF

DEBFW
Values that can be written to the field DEBF