Module atsamd51j19a::sdhc0::bgcr

Block Gap Control

Structs

R
Value read from the register

W
Value to write to the register

_CONTRW
Proxy

_INTBGW
Proxy

_RWCTRLW
Proxy

_STPBGRW
Proxy

Enums

CONTRR
Possible values of the field CONTR

CONTRW
Values that can be written to the field CONTR

INTBGR
Possible values of the field INTBG

INTBGW
Values that can be written to the field INTBG

RWCTRLR
Possible values of the field RWCTRL

RWCTRLW
Values that can be written to the field RWCTRL

STPBGRR
Possible values of the field STPBGR

STPBGRW
Values that can be written to the field STPBGR