Enum atsamd51j19a::sdhc0::ca1r::TCNTRTR

pub enum TCNTRTR {
    DISABLED,
    _1S,
    _2S,
    _4S,
    _8S,
    _16S,
    _32S,
    _64S,
    _128S,
    _256S,
    _512S,
    _1024S,
    OTHER,
    _Reserved(u8),
}

Possible values of the field TCNTRT

Variants
DISABLED

Re-Tuning Timer disabled

_1S

1 second

_2S

2 seconds

_4S

4 seconds

_8S

8 seconds

_16S

16 seconds

_32S

32 seconds

_64S

64 seconds

_128S

128 seconds

_256S

256 seconds

_512S

512 seconds

_1024S

1024 seconds

OTHER

Get information from other source

_Reserved(u8)

Reserved

Methods

impl TCNTRTR

pub fn bits(&self) -> u8

Value of the field as raw bits

pub fn is_disabled(&self) -> bool

Checks if the value of the field is DISABLED

pub fn is_1s(&self) -> bool

Checks if the value of the field is _1S

pub fn is_2s(&self) -> bool

Checks if the value of the field is _2S

pub fn is_4s(&self) -> bool

Checks if the value of the field is _4S

pub fn is_8s(&self) -> bool

Checks if the value of the field is _8S

pub fn is_16s(&self) -> bool

Checks if the value of the field is _16S

pub fn is_32s(&self) -> bool

Checks if the value of the field is _32S

pub fn is_64s(&self) -> bool

Checks if the value of the field is _64S

pub fn is_128s(&self) -> bool

Checks if the value of the field is _128S

pub fn is_256s(&self) -> bool

Checks if the value of the field is _256S

pub fn is_512s(&self) -> bool

Checks if the value of the field is _512S

pub fn is_1024s(&self) -> bool

Checks if the value of the field is _1024S

pub fn is_other(&self) -> bool

Checks if the value of the field is OTHER