Module atsamd51j19a::sdhc0::eisier

Error Interrupt Signal Enable

Structs

R
Value read from the register

W
Value to write to the register

_ACMDW
Proxy

_ADMAW
Proxy

_CMDCRCW
Proxy

_CMDENDW
Proxy

_CMDIDXW
Proxy

_CMDTEOW
Proxy

_CURLIMW
Proxy

_DATCRCW
Proxy

_DATENDW
Proxy

_DATTEOW
Proxy

Enums

ACMDR
Possible values of the field ACMD

ACMDW
Values that can be written to the field ACMD

ADMAR
Possible values of the field ADMA

ADMAW
Values that can be written to the field ADMA

CMDCRCR
Possible values of the field CMDCRC

CMDCRCW
Values that can be written to the field CMDCRC

CMDENDR
Possible values of the field CMDEND

CMDENDW
Values that can be written to the field CMDEND

CMDIDXR
Possible values of the field CMDIDX

CMDIDXW
Values that can be written to the field CMDIDX

CMDTEOR
Possible values of the field CMDTEO

CMDTEOW
Values that can be written to the field CMDTEO

CURLIMR
Possible values of the field CURLIM

CURLIMW
Values that can be written to the field CURLIM

DATCRCR
Possible values of the field DATCRC

DATCRCW
Values that can be written to the field DATCRC

DATENDR
Possible values of the field DATEND

DATENDW
Values that can be written to the field DATEND

DATTEOR
Possible values of the field DATTEO

DATTEOW
Values that can be written to the field DATTEO