Module atsamd51j19a::sdhc0::nister

Normal Interrupt Status Enable

Structs

R
Value read from the register

W
Value to write to the register

_BLKGEW
Proxy

_BRDRDYW
Proxy

_BWRRDYW
Proxy

_CINSW
Proxy

_CINTW
Proxy

_CMDCW
Proxy

_CREMW
Proxy

_DMAINTW
Proxy

_TRFCW
Proxy

Enums

BLKGER
Possible values of the field BLKGE

BLKGEW
Values that can be written to the field BLKGE

BRDRDYR
Possible values of the field BRDRDY

BRDRDYW
Values that can be written to the field BRDRDY

BWRRDYR
Possible values of the field BWRRDY

BWRRDYW
Values that can be written to the field BWRRDY

CINSR
Possible values of the field CINS

CINSW
Values that can be written to the field CINS

CINTR
Possible values of the field CINT

CINTW
Values that can be written to the field CINT

CMDCR
Possible values of the field CMDC

CMDCW
Values that can be written to the field CMDC

CREMR
Possible values of the field CREM

CREMW
Values that can be written to the field CREM

DMAINTR
Possible values of the field DMAINT

DMAINTW
Values that can be written to the field DMAINT

TRFCR
Possible values of the field TRFC

TRFCW
Values that can be written to the field TRFC