Enum atsamd51j19a::wdt::config::PERR

pub enum PERR {
    CYC8,
    CYC16,
    CYC32,
    CYC64,
    CYC128,
    CYC256,
    CYC512,
    CYC1024,
    CYC2048,
    CYC4096,
    CYC8192,
    CYC16384,
    _Reserved(u8),
}

Possible values of the field PER

Variants
CYC8

8 clock cycles

CYC16

16 clock cycles

CYC32

32 clock cycles

CYC64

64 clock cycles

CYC128

128 clock cycles

CYC256

256 clock cycles

CYC512

512 clock cycles

CYC1024

1024 clock cycles

CYC2048

2048 clock cycles

CYC4096

4096 clock cycles

CYC8192

8192 clock cycles

CYC16384

16384 clock cycles

_Reserved(u8)

Reserved

Methods

impl PERR

pub fn bits(&self) -> u8

Value of the field as raw bits

pub fn is_cyc8(&self) -> bool

Checks if the value of the field is CYC8

pub fn is_cyc16(&self) -> bool

Checks if the value of the field is CYC16

pub fn is_cyc32(&self) -> bool

Checks if the value of the field is CYC32

pub fn is_cyc64(&self) -> bool

Checks if the value of the field is CYC64

pub fn is_cyc128(&self) -> bool

Checks if the value of the field is CYC128

pub fn is_cyc256(&self) -> bool

Checks if the value of the field is CYC256

pub fn is_cyc512(&self) -> bool

Checks if the value of the field is CYC512

pub fn is_cyc1024(&self) -> bool

Checks if the value of the field is CYC1024

pub fn is_cyc2048(&self) -> bool

Checks if the value of the field is CYC2048

pub fn is_cyc4096(&self) -> bool

Checks if the value of the field is CYC4096

pub fn is_cyc8192(&self) -> bool

Checks if the value of the field is CYC8192

pub fn is_cyc16384(&self) -> bool

Checks if the value of the field is CYC16384